+++ src/Schema/SchemaCatalog.cs
@@ -14,15 +14,16 @@ namespace EditorConfig
         /// <summary>The name of the root keyword.</summary>
         public const string Root = "root";
 
-        private static IEnumerable<Keyword> _allKeywords;
-
         static SchemaCatalog()
         {
             ParseJson();
         }
 
+        /// <summary>A list of all keywords including the ones marked as hidden.</summary>
+        public static IEnumerable<Keyword> AllKeywords { get; private set; }
+
         /// <summary>A list of all visible keywords.</summary>
-        public static IEnumerable<Keyword> Keywords { get; private set; }
+        public static IEnumerable<Keyword> VisibleKeywords { get; private set; }
 
         /// <summary>A list of all severities.</summary>
         public static IEnumerable<Severity> Severities { get; private set; }
@@ -30,7 +31,7 @@ namespace EditorConfig
         /// <summary>Tries to get a keyword by name.</summary>
         public static bool TryGetKeyword(string name, out Keyword keyword)
         {
-            keyword = _allKeywords.FirstOrDefault(c => c.Name.Is(name));
+            keyword = AllKeywords.FirstOrDefault(c => c.Name.Is(name));
 
             return keyword != null;
         }
@@ -51,8 +52,8 @@ namespace EditorConfig
             var obj = JObject.Parse(File.ReadAllText(file));
 
             Severities = JsonConvert.DeserializeObject<IEnumerable<Severity>>(obj["severities"].ToString());
-            _allKeywords = JsonConvert.DeserializeObject<IEnumerable<Keyword>>(obj["properties"].ToString());
-            Keywords = _allKeywords.Where(p => p.IsVisible);
+            AllKeywords = JsonConvert.DeserializeObject<IEnumerable<Keyword>>(obj["properties"].ToString());
+            VisibleKeywords = AllKeywords.Where(p => p.IsVisible);
         }
     }
 }